requirements.txt
aiofiles==22.1.0
PyYAML==6.0
requests==2.28.1
Telethon==1.26.1
aiohttp~=3.8.3